+++ src/curvefit/models/core_model.py
@@ -89,7 +89,8 @@ class CoreModel(Model):
     def convert_inputs(self, data):
         if isinstance(data, DataInputs):
             self.data_inputs = data
-            
+            return 
+
         df = data[0]
         data_specs = data[1]
